Debug controller add cmdline decoder Remove RandomConstructive from autoconfig
